shop paged finished for now added styling and base layout
