+++ src/components/About.jsx
@@ -1,0 +1,14 @@
+const About = () => {
+  return (
+    <>
+      <header>
+        {" "}
+        {/* //? Banner for background img */}
+        <h1>About</h1>
+        <p>""In every drawing there's a story""</p>2
+      </header>
+    </>
+  );
+};
+
+export default About;
